상점 화면 구성 버튼 이벤트 아직 X
