+++ routie-mobile/app/src/main/res/values/colors.xml
@@ -2,4 +2,6 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="background_color">#F5F1EB</color>
+    <color name="font_color">#453536</color>
 </resources>
